refactor(#1682): delete mcp_client.py + mcp_server.py back-compat shims
Both were pure #4-era re-export shims (mcp_client.py 11L, mcp_server.py
16L) forwarding to reyn.mcp.client / reyn.mcp.server (the transport core,
which correctly stays top-level — A3b mcp/→tools was cancelled).

REWRITE (3-class × all importing dirs, incl tests):
- Form A — dotted: reyn.mcp_client → reyn.mcp.client, reyn.mcp_server →
  reyn.mcp.server (24 files, incl the test-private-internal importers
  `from reyn.mcp_server import _MCPProgressBridge / _MCPInterventionBus /
  _get_agent_lock` — those internals originate in reyn.mcp.server).
- Form B — bare: `from reyn import mcp_server` → `from reyn.mcp import
  server as mcp_server` (2 test files; preserves mcp_server.X usage).
- Then deleted both shim files.

Verified: 0 dotted reyn.mcp_{client,server} + 0 form-B remain; pyproject /
reyn.__init__ clean; the plural mcp_servers/mcp_clients config fields are
unrelated and untouched. Advances #7 (2 flat modules eliminated) + removes
#4 cruft per the no-lingering-shim precedent.
