+++ tests/test_mcp_progress_and_timeout.py
@@ -36,7 +36,7 @@ import pytest
 from reyn.chat.forwarder import ChatEventForwarder
 from reyn.chat.outbox import OutboxMessage
 from reyn.core.events.events import EventLog
-from reyn.mcp_client import MCPClient
+from reyn.mcp.client import MCPClient
 from reyn.schemas.models import MCPIROp
 
 # ── 1. MCPClient.call_tool signature accepts the new kwargs ────────────
