+++ tests/test_mcp_server_progress_271.py
@@ -45,7 +45,7 @@ import pytest
 pytest.importorskip("mcp", reason="mcp not installed")
 
 from reyn.core.events.events import EventLog  # noqa: E402
-from reyn.mcp_server import _MCPProgressBridge  # noqa: E402
+from reyn.mcp.server import _MCPProgressBridge  # noqa: E402
 from reyn.schemas.models import Event  # noqa: E402
