+++ tests/test_transport_ref.py
@@ -468,7 +468,7 @@ async def test_a2a_endpoint_uses_message_bus(tmp_path, monkeypatch):
     """
     from reyn.chat.profile import AgentProfile
     from reyn.chat.registry import AgentRegistry
-    from reyn.mcp_server import send_to_agent_impl
+    from reyn.mcp.server import send_to_agent_impl
     from reyn.runtime.budget.budget import BudgetTracker, CostConfig
 
     state_log = StateLog(tmp_path / ".reyn" / "state" / "wal.jsonl")
